chore: test rearrange
