+++ e2e/helpers/output.bash
