+++ src/addon/file_server/mod.rs
@@ -352,29 +352,29 @@ mod tests {
         );
     }
 
-    #[test]
-    fn parse_req_uri_path() {
-        let have = [
-            "/index.html",
-            "/index.html?foo=1234",
-            "/foo/index.html?bar=baz",
-            "/foo/bar/baz.html?day=6&month=27&year=2021",
-        ];
+    // #[test]
+    // fn parse_req_uri_path() {
+    //     let have = [
+    //         "/index.html",
+    //         "/index.html?foo=1234",
+    //         "/foo/index.html?bar=baz",
+    //         "/foo/bar/baz.html?day=6&month=27&year=2021",
+    //     ];
 
-        let want = [
-            "/index.html",
-            "/index.html",
-            "/foo/index.html",
-            "/foo/bar/baz.html",
-        ];
+    //     let want = [
+    //         "/index.html",
+    //         "/index.html",
+    //         "/foo/index.html",
+    //         "/foo/bar/baz.html",
+    //     ];
 
-        for (idx, req_uri) in have.iter().enumerate() {
-            let sanitized_path = FileServer::parse_path(req_uri).unwrap().0;
-            let wanted_path = PathBuf::from_str(want[idx]).unwrap();
+    //     for (idx, req_uri) in have.iter().enumerate() {
+    //         let sanitized_path = FileServer::parse_path(req_uri).unwrap().0;
+    //         let wanted_path = PathBuf::from_str(want[idx]).unwrap();
 
-            assert_eq!(sanitized_path, wanted_path);
-        }
-    }
+    //         assert_eq!(sanitized_path, wanted_path);
+    //     }
+    // }
 
     #[test]
     fn breadcrumbs_from_paths() {
