+++ src/lib.rs
@@ -4,6 +4,9 @@ mod config;
 mod server;
 mod utils;
 
+#[cfg(test)]
+mod test;
+
 use anyhow::{Context, Result};
 use std::convert::TryFrom;
 use structopt::StructOpt;
