+++ src/test/mod.rs
@@ -1,3 +1,4 @@
 mod basic_auth;
 mod cors;
 mod defacto;
+mod gzip;
